update imports and move addfilesform to claim-files-tab folder
